requirements.txt
alpaca-py==0.13.1
annotated-types==0.6.0
certifi==2023.7.22
charset-normalizer==3.3.1
idna==3.4
msgpack==1.0.7
numpy==1.26.1
pandas==2.1.2
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sseclient-py==1.8.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
websockets==11.0.3
